add future plans
we are giving the "next-meeting' link to sister groups, so expect more action for this template.  thus it would be great to say everything we know in order to pique visitor interest.  we can hardcode these two lists here, or if you prefer, we can pull the lists from two tables.  my thinking is that things are early and fluid enough that the easier hardcoding route is fine for now.  or an intermediate route could be to define a future_plans html slug, cuz that would be useful in the monthly meeting announcement as well as here [assuming we can pull such a slug into the monthly "newsletter"].